==> evaluations/reads/2-create-input-list.ipynb <==
# 1. Load metadata

import pandas as pd

campy_df = pd.read_csv('data/datasets/datasets/Campylobacter_jejuni_0810PADBR-1.tsv', sep='\t', skiprows=8)
campy_df = campy_df[~campy_df['strain'].isna()]
campy_df = campy_df.sample(frac=1, random_state=97)
print(len(campy_df))
campy_df.head()

ecoli_df = pd.read_csv('data/datasets/datasets/Escherichia_coli_1405WAEXK-1.tsv', sep='\t', skiprows=7)
ecoli_df = ecoli_df[~ecoli_df['strain'].isna()]
ecoli_df = ecoli_df.sample(frac=1, random_state=97)
ecoli_df.head()

listeria_df = pd.read_csv('data/datasets/datasets/Listeria_monocytogenes_1408MLGX6-3WGS.tsv', sep='\t', skiprows=8)
listeria_df = listeria_df[~listeria_df['strain'].isna()]
listeria_df = listeria_df.sample(frac=1, random_state=97)
listeria_df.head()

salmonella_df = pd.read_csv('data/datasets/datasets/Salmonella_enterica_1203NYJAP-1.tsv', sep='\t', skiprows=7)
salmonella_df = salmonella_df[~salmonella_df['strain'].isna()]
salmonella_df = salmonella_df.sample(frac=1, random_state=97)
salmonella_df.head()

metadata_df = pd.concat([campy_df, ecoli_df, listeria_df, salmonella_df])
metadata_df['strain'] = metadata_df['strain'].str.strip()
metadata_df = metadata_df[metadata_df['SRArun_acc'] != '-']

columns_to_keep = ['dataSetName', 'strain', 'outbreak', 'SRArun_acc']

metadata_df = metadata_df.loc[:, columns_to_keep].reset_index(drop=True)
outgroup_df = metadata_df.loc[metadata_df['outbreak'] == 'outgroup', :]
metadata_df.loc[outgroup_df.index, 'outbreak'] = outgroup_df.apply(lambda x: x['outbreak'] + '_' + x['dataSetName'], axis='columns')

metadata_df.loc[metadata_df['dataSetName'] == '0810PADBR-1', 'reference'] = 'GCA_001879185.2_ASM187918v2_genomic'
metadata_df.loc[metadata_df['dataSetName'] == '0810PADBR-1', 'Organism'] = 'Campylobacter jejuni'
metadata_df.loc[metadata_df['dataSetName'] == '1405WAEXK-1', 'reference'] = 'GCF_000703365.1_Ec2011C-3609_genomic'
metadata_df.loc[metadata_df['dataSetName'] == '1405WAEXK-1', 'Organism'] = 'Escherichia coli'
metadata_df.loc[metadata_df['dataSetName'] == '1408MLGX6-3WGS', 'reference'] = 'GCF_001047715.2_ASM104771v2_genomic'
metadata_df.loc[metadata_df['dataSetName'] == '1408MLGX6-3WGS', 'Organism'] = 'Listeria monocytogenes'
metadata_df.loc[metadata_df['dataSetName'] == '1203NYJAP-1 - Tuna Scrape Outbreak', 'reference'] = 'GCF_000439415.1_ASM43941v1_genomic'
metadata_df.loc[metadata_df['dataSetName'] == '1203NYJAP-1 - Tuna Scrape Outbreak', 'Organism'] = 'Salmonella enterica'

metadata_df

metadata_df.to_csv('metadata.tsv', sep='\t', index=False)
