==> evaluations/reads/4-read-stats.ipynb <==
# 1. Parameters

from pathlib import Path
import pandas as pd
import io

dataset_input_reads = {
    '0810PADBR-1': Path('cases/full/0810PADBR-1_input-files-case.tsv'),
    '1405WAEXK-1': Path('cases/full/1405WAEXK-1_input-files-case.tsv'),
    '1408MLGX6-3WGS': Path('cases/full/1408MLGX6-3WGS_input-files-case.tsv'),
    '1203NYJAP-1 - Tuna Scrape Outbreak': Path('cases/full/1203NYJAP-1_-_Tuna_Scrape_Outbreak_input-files-case.tsv')
}
dataset_input_reads

# 2. Reads datasets input files

dataset_input_dfs = {}

for dataset in dataset_input_reads:
    input_df = pd.read_csv(dataset_input_reads[dataset], sep='\t')
    
    dataset_input_dfs[dataset] = input_df.set_index('Sample')
    
dataset_input_dfs['1405WAEXK-1']

# 3. Get fastq stats

def read_dataset_stats(dataset: str) -> pd.DataFrame:
    reads1_list = dataset_input_dfs[dataset]['Reads1'].tolist()
    reads2_list = dataset_input_dfs[dataset]['Reads2'].tolist()
    files = ' '.join(reads1_list + reads2_list)
    stats_str = !seqkit stats --threads 32 --basename --tabular {files}

    dataset_df = pd.read_csv(io.StringIO('\n'.join(stats_str)), sep='\t')
    dataset_df['Dataset'] = dataset
    dataset_df['Sample'] = dataset_df['file'].str.replace('_[12].fastq.gz$', '', regex=True)
    grouped_df = dataset_df.groupby('Sample').agg({'Dataset': 'first', 'num_seqs': 'sum', 'sum_len': 'sum'})
    grouped_df = grouped_df.rename({'num_seqs': 'Number reads', 'sum_len': 'Number nucleotides'}, axis='columns')
    return grouped_df

dataset_dfs = []
for dataset in dataset_input_dfs:
    dataset_df = read_dataset_stats(dataset)
    dataset_dfs.append(dataset_df)

read_stats_df = pd.concat(dataset_dfs)
read_stats_df

# 4. Save fastq stats

read_stats_df.to_csv('cases/full/reads-stats.tsv', sep='\t')
